+++ screens/Journal.js
@@ -54,19 +54,19 @@ export default class JournalScreen extends React.Component {
           exercise: 7.7,
           happiness: 4,
           activity: 3,
-          date: new Date()
+          date: new Date('2019-03-07T03:24:00')
         },
         {
           exercise: 2.3,
           happiness: 4,
           activity: 3,
-          date: new Date()
+          date: new Date('2019-03-06T03:24:00')
         },
         {
           exercise: 4.4,
           happiness: 4,
           activity: 3,
-          date: new Date()
+          date: new Date('2019-03-05T03:24:00')
         }
       ]
     });